+++ server/src/main/java/com/blocklang/config/OAuth2LoginSecurityConfig.java
@@ -1,28 +1,22 @@
 package com.blocklang.config;
 
-import org.springframework.boot.autoconfigure.security.oauth2.client.EnableOAuth2Sso;
-import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
 
-@Configuration
-@EnableOAuth2Sso
-public class SecurityConfig extends WebSecurityConfigurerAdapter {
+@EnableWebSecurity
+public class OAuth2LoginSecurityConfig extends WebSecurityConfigurerAdapter{
 
 	@Override
 	protected void configure(HttpSecurity http) throws Exception {
 		// We recommend disabling CSRF protection completely only if you are creating a
 		// service that is used by non-browser clients
 		// TODO: 考虑通过将服务拆分到单独项目中，然后打开此功能
-		http.csrf().disable(); // 不加此段的话，测试用例会报 403
+		http.csrf().disable();
 		
-		http.antMatcher("/**")
-			.authorizeRequests()
-				.antMatchers(Resources.PUBLIC_URL)
-				.permitAll()
-			.anyRequest()
-				.authenticated();
-		
+		http
+			.oauth2Login()
+				.loginPage("/"); // 因为是 single page app，所以将登录页设置为首页
 	}
 
 }