+++ server/src/test/java/com/blocklang/core/test/AbstractServiceTest.java
@@ -1,4 +1,4 @@
-package com.blocklang.core.service;
+package com.blocklang.core.test;
 
 import javax.transaction.Transactional;