+++ server/src/test/java/com/blocklang/release/service/impl/AppServiceImplTest.java
@@ -9,7 +9,7 @@ import java.util.Optional;
 import org.junit.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 
-import com.blocklang.core.service.AbstractServiceTest;
+import com.blocklang.core.test.AbstractServiceTest;
 import com.blocklang.release.dao.AppDao;
 import com.blocklang.release.model.App;
 import com.blocklang.release.service.AppService;