+++ server/src/test/java/com/blocklang/release/service/impl/BuildServiceImplTest.java
@@ -6,7 +6,7 @@ import java.time.LocalDateTime;
 import org.junit.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 
-import com.blocklang.core.service.AbstractServiceTest;
+import com.blocklang.core.test.AbstractServiceTest;
 import com.blocklang.develop.model.Project;
 import com.blocklang.release.constant.ReleaseResult;
 import com.blocklang.release.dao.AppDao;