+++ server/src/test/java/com/blocklang/release/service/impl/ProjectReleaseTaskServiceImplTest.java
@@ -9,7 +9,7 @@ import java.time.LocalDateTime;
 import org.junit.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 
-import com.blocklang.core.service.AbstractServiceTest;
+import com.blocklang.core.test.AbstractServiceTest;
 import com.blocklang.release.constant.ReleaseResult;
 import com.blocklang.release.model.ProjectReleaseTask;
 import com.blocklang.release.service.ProjectReleaseTaskService;